moved NoteContent class to separate file
