+++ lib/models/note_content.dart
@@ -1,0 +1,89 @@
+/// Data models for notes
+///
+/// Copyright (C) 2023-2025, Software Innovation Institute
+///
+/// Licensed under the GNU General Public License, Version 3 (the "License");
+///
+/// License: https://opensource.org/license/gpl-3-0
+//
+// Time-stamp: <Thursday 2026-03-13 17:53:12 +1100 Graham Williams>
+//
+// This program is free software: you can redistribute it and/or modify it under
+// the terms of the GNU General Public License as published by the Free Software
+// Foundation, either version 3 of the License, or (at your option) any later
+// version.
+//
+// This program is distributed in the hope that it will be useful, but WITHOUT
+// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+// FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+// details.
+//
+// You should have received a copy of the GNU General Public License along with
+// this program.  If not, see <https://opensource.org/license/gpl-3-0>.
+///
+/// Authors: Jess Moore
+
+library;
+
+import 'package:solidpod/solidpod.dart';
+
+import 'package:notepod/constants/turtle_structures.dart';
+
+/// Base data model for the nested note within a note object
+
+class NoteContent {
+  final String noteTitle;
+  final String createdDateTime;
+  final String modifiedDateTime;
+  final String noteContent;
+  final List<String> authUsers;
+
+  const NoteContent({
+    required this.noteTitle,
+    required this.createdDateTime,
+    required this.modifiedDateTime,
+    required this.noteContent,
+    this.authUsers = const [],
+  });
+
+  /// Method to create NoteContent object from json data map
+
+  factory NoteContent.fromJson(Map<String, dynamic> json) {
+    return NoteContent(
+      noteTitle: json[noteTitlePred] as String,
+      createdDateTime: json[createdDateTimePred] as String,
+      modifiedDateTime: json[modifiedDateTimePred] as String,
+      noteContent: json[noteContentPred] as String,
+      authUsers: (json[authUserPred] as Map).keys.toList().cast<String>(),
+    );
+  }
+
+  /// Method to export NoteContent object to json data map
+
+  Map<String, dynamic> toJson() => {
+        noteTitlePred: noteTitle,
+        createdDateTimePred: createdDateTime,
+        modifiedDateTimePred: modifiedDateTime,
+        noteContentPred: noteContent,
+        authUserPred: authUsers,
+      };
+
+  /// Copy method for creating a new instance that is an
+  /// updated copy of another instance
+
+  NoteContent copyWith({
+    String? noteTitle,
+    String? createdDateTime,
+    String? modifiedDateTime,
+    String? noteContent,
+    List<String>? authUsers,
+  }) {
+    return NoteContent(
+      noteTitle: noteTitle ?? this.noteTitle,
+      createdDateTime: createdDateTime ?? this.createdDateTime,
+      modifiedDateTime: modifiedDateTime ?? this.modifiedDateTime,
+      noteContent: noteContent ?? this.noteContent,
+      authUsers: authUsers ?? this.authUsers,
+    );
+  }
+}
