+++ lib/models/own_note.dart
@@ -28,7 +28,7 @@ library;
 import 'package:solidpod/solidpod.dart';
 
 import 'package:notepod/constants/turtle_structures.dart';
-import 'package:notepod/models/note.dart';
+import 'package:notepod/models/note_content.dart';
 
 /// Data model for user's note
